Move player rendering to gameLoop
